Rewrite: Move Set/Map.{+,-} and Map.zip to Experimental rules.
Terms don't have type in scalameta yet: https://github.com/scalameta/scalameta/issues/1212

The SymbolMatcher is not enought for those rules since we need to check the type of the lhs.
One workarround is to jump to the symbol if we see an identifier on the lhs. for example:

iset: immutable.Set[Int]
// ...
iset + 1

we can jump to the symbol of iset and get it's type.

This is really fragile, since it's not going to work if we have aliases/subtyping.

However, it's useful to create those rules, since when we are unblock by scalameta#1212 it's
trivial to add the remaining piece
